more tidying up
