+++ handlers/sf-contact-merge/src/main/scala/com/gu/sf_contact_merge/validate/ValidationSteps.scala
@@ -1,7 +1,7 @@
 package com.gu.sf_contact_merge.validate
 
-import com.gu.sf_contact_merge.validate.GetContacts.IdentityId
-import com.gu.sf_contact_merge.validate.GetIdentityAndZuoraEmailsForAccounts.IdentityAndSFContactAndEmail
+import com.gu.sf_contact_merge.getaccounts.GetContacts.IdentityId
+import com.gu.sf_contact_merge.getaccounts.GetIdentityAndZuoraEmailsForAccountsSteps.IdentityAndSFContactAndEmail
 import com.gu.util.apigateway.ApiGatewayResponse
 import com.gu.util.reader.Types._
