+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/update/UpdateAccountSFLinksTest.scala
@@ -1,7 +1,7 @@
 package com.gu.sf_contact_merge.update
 
 import com.gu.sf_contact_merge.update.UpdateAccountSFLinks.{CRMAccountId, LinksFromZuora}
-import com.gu.sf_contact_merge.validate.GetContacts.{AccountId, IdentityId, SFContactId}
+import com.gu.sf_contact_merge.getaccounts.GetContacts.{AccountId, IdentityId, SFContactId}
 import com.gu.util.resthttp.RestRequestMaker.{PutRequest, RelativePath}
 import org.scalatest.{FlatSpec, Matchers}
 import play.api.libs.json._
