+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/validate/AssertSameEmailsTest.scala
@@ -1,6 +1,6 @@
 package com.gu.sf_contact_merge.validate
 
-import com.gu.sf_contact_merge.validate.GetEmails.EmailAddress
+import com.gu.sf_contact_merge.getaccounts.GetEmails.EmailAddress
 import org.scalatest.{FlatSpec, Matchers}
 
 class AssertSameEmailsTest extends FlatSpec with Matchers {
