+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/validate/GetContactsTest.scala
@@ -1,7 +1,8 @@
 package com.gu.sf_contact_merge.validate
 
-import com.gu.sf_contact_merge.validate.GetContacts.{IdentityAndSFContact, AccountId, IdentityId, SFContactId}
-import com.gu.sf_contact_merge.validate.GetEmails.ContactId
+import com.gu.sf_contact_merge.getaccounts.GetContacts
+import com.gu.sf_contact_merge.getaccounts.GetContacts.{AccountId, IdentityAndSFContact, IdentityId, SFContactId}
+import com.gu.sf_contact_merge.getaccounts.GetEmails.ContactId
 import com.gu.util.resthttp.Types.ClientSuccess
 import com.gu.zuora.fake.FakeZuoraQuerier
 import org.scalatest.{FlatSpec, Matchers}
