+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/validate/GetEmailsTest.scala
@@ -1,6 +1,7 @@
 package com.gu.sf_contact_merge.validate
 
-import com.gu.sf_contact_merge.validate.GetEmails.{ContactId, EmailAddress}
+import com.gu.sf_contact_merge.getaccounts.GetEmails
+import com.gu.sf_contact_merge.getaccounts.GetEmails.{ContactId, EmailAddress}
 import com.gu.util.resthttp.Types.ClientSuccess
 import com.gu.zuora.fake.FakeZuoraQuerier
 import org.scalatest.{FlatSpec, Matchers}
